+++ gazon-service/src/main/java/com/gitlab/controller/SearchProductController.java
@@ -13,14 +13,16 @@ import java.util.List;
 @RequiredArgsConstructor
 public class SearchProductController implements SearchProductRestApi {
 
+
     private final ProductService productService;
 
     @Override
-    public ResponseEntity<List<ProductDto>> search(String name) {
-        var products = productService.findByNameIgnoreCaseContaining(name);
+    public ResponseEntity<List<ProductDto>> search(String name) throws InterruptedException {
+        List<ProductDto> products = productService.findByNameIgnoreCaseContaining(name);
 
-        return products.isEmpty() ?
-                ResponseEntity.noContent().build() :
-                ResponseEntity.ok(products);
+
+        return !products.isEmpty() ?
+                ResponseEntity.ok(products) :
+        ResponseEntity.noContent().build();
     }
 }