+++ security-service/src/main/java/com/gitlab/dto/AuthRequest.java
@@ -1,0 +1,14 @@
+package com.gitlab.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class AuthRequest {
+
+    private String email;
+    private String password;
+}