+++ security-service/src/main/java/com/gitlab/model/User.java
@@ -1,0 +1,100 @@
+package com.gitlab.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.JoinTable;
+import javax.persistence.NamedAttributeNode;
+import javax.persistence.NamedEntityGraph;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import java.time.LocalDate;
+import java.util.Collection;
+import java.util.Set;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "security_users", schema = "public", catalog = "postgres")
+@NamedEntityGraph(name = "userWithSets",
+        attributeNodes = {@NamedAttributeNode("rolesSet")})
+public class User implements UserDetails {
+
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    private Long id;
+
+    @Column(name = "email")
+    private String email;
+
+    @Column(name = "password")
+    private String password;
+
+    @Column(name = "security_question")
+    private String securityQuestion;
+
+    @Column(name = "answer_question")
+    private String answerQuestion;
+
+    @Column(name = "first_name")
+    private String firstName;
+
+    @Column(name = "last_name")
+    private String lastName;
+
+    @Column(name = "create_date")
+    private LocalDate createDate;
+
+    @OneToMany(cascade = CascadeType.ALL)
+    @JoinTable(name = "security_users_roles",
+            joinColumns = @JoinColumn(name = "security_user_id"),
+            inverseJoinColumns = @JoinColumn(name = "security_role_id"))
+    private Set<Role> rolesSet;
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return rolesSet;
+    }
+
+    @Override
+    public String getUsername() {
+        return email;
+    }
+
+    @Override
+    public String getPassword() {
+        return password;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+
+}