+++ security-service/src/main/java/com/gitlab/repository/RoleRepository.java
@@ -1,0 +1,13 @@
+package com.gitlab.repository;
+
+import com.gitlab.model.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface RoleRepository extends JpaRepository<Role, Long> {
+
+    Role findByName (String name);
+
+}